More stuff in product-core
Misc

More lazy AboutDialog

Only output version in product-core

Split product-core into embedded-core/app-core
