+++ packages/app-core/src/version.js
@@ -1,0 +1,1 @@
+export const version = '2.5.0'
