+++ packages/embedded-core/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@jbrowse/product-core",
+  "name": "@jbrowse/embedded-core",
   "version": "2.5.0",
-  "description": "JBrowse 2 code shared between products but not used by plugins",
+  "description": "JBrowse 2 code shared between embedded products",
   "keywords": [
     "jbrowse",
     "jbrowse2",
@@ -15,7 +15,7 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/GMOD/jbrowse-components.git",
-    "directory": "packages/product-core"
+    "directory": "packages/embedded-core"
   },
   "author": "JBrowse Team",
   "distMain": "dist/index.js",
@@ -33,7 +33,7 @@
     "build:esm": "tsc --build tsconfig.build.esm.json",
     "build:es5": "tsc --build tsconfig.build.es5.json",
     "build": "npm run build:esm && npm run build:es5",
-    "test": "cd ../..; jest packages/product-core",
+    "test": "cd ../..; jest packages/text-indexing",
     "clean": "rimraf dist esm *.tsbuildinfo",
     "prebuild": "yarn clean",
     "prepack": "yarn build && yarn useDist",
@@ -44,6 +44,7 @@
   "dependencies": {
     "@babel/runtime": "^7.16.3",
     "@mui/material": "^5.10.17",
+    "@jbrowse/product-core": "^2.5.0",
     "copy-to-clipboard": "^3.3.1",
     "react-error-boundary": "^4.0.3",
     "shortid": "^2.2.15"
