+++ packages/product-core/src/index.ts
@@ -1,3 +1,4 @@
 export * as RootModel from './RootModel'
 export * as Session from './Session'
 export { isSession } from './Session'
+export { default as AboutDialog } from './ui/AboutDialog'
