+++ products/jbrowse-desktop/src/sessionModel/AboutDialog.tsx
@@ -1,0 +1,2 @@
+// default re-export for react lazy
+export { AboutDialog as default } from '@jbrowse/product-core'
