+++ products/jbrowse-react-linear-genome-view/src/JBrowseLinearGenomeView/JBrowseLinearGenomeView.tsx
@@ -3,12 +3,11 @@ import { observer } from 'mobx-react'
 import { getEnv } from 'mobx-state-tree'
 import { readConfObject } from '@jbrowse/core/configuration'
 import { createJBrowseTheme } from '@jbrowse/core/ui'
+import { EmbeddedViewContainer, ModalWidget } from '@jbrowse/embedded-core'
 import { ThemeProvider, ScopedCssBaseline } from '@mui/material'
 import { makeStyles } from 'tss-react/mui'
 
 // locals
-import ModalWidget from './ModalWidget'
-import ViewContainer from './ViewContainer'
 import { ViewModel } from '../createModel/createModel'
 
 const useStyles = makeStyles()(() => ({
@@ -19,36 +18,38 @@ const useStyles = makeStyles()(() => ({
   },
 }))
 
-const JBrowseLinearGenomeView = observer(
-  ({ viewState }: { viewState: ViewModel }) => {
-    const { classes } = useStyles()
-    const { session } = viewState
-    const { view } = session
-    const { pluginManager } = getEnv(session)
-    const viewType = pluginManager.getViewType(view.type)
-    if (!viewType) {
-      throw new Error(`unknown view type ${view.type}`)
-    }
-    const { ReactComponent } = viewType
-    const theme = createJBrowseTheme(
-      readConfObject(viewState.config.configuration, 'theme'),
-    )
+const JBrowseLinearGenomeView = observer(function ({
+  viewState,
+}: {
+  viewState: ViewModel
+}) {
+  const { classes } = useStyles()
+  const { session } = viewState
+  const { view } = session
+  const { pluginManager } = getEnv(session)
+  const viewType = pluginManager.getViewType(view.type)
+  if (!viewType) {
+    throw new Error(`unknown view type ${view.type}`)
+  }
+  const { ReactComponent } = viewType
+  const theme = createJBrowseTheme(
+    readConfObject(viewState.config.configuration, 'theme'),
+  )
 
-    return (
-      <ThemeProvider theme={theme}>
-        <div className={classes.avoidParentStyle}>
-          <ScopedCssBaseline>
-            <ViewContainer key={`view-${view.id}`} view={view}>
-              <Suspense fallback={<div>Loading...</div>}>
-                <ReactComponent model={view} session={session} />
-              </Suspense>
-            </ViewContainer>
-            <ModalWidget session={session} />
-          </ScopedCssBaseline>
-        </div>
-      </ThemeProvider>
-    )
-  },
-)
+  return (
+    <ThemeProvider theme={theme}>
+      <div className={classes.avoidParentStyle}>
+        <ScopedCssBaseline>
+          <EmbeddedViewContainer key={`view-${view.id}`} view={view}>
+            <Suspense fallback={<div>Loading...</div>}>
+              <ReactComponent model={view} session={session} />
+            </Suspense>
+          </EmbeddedViewContainer>
+          <ModalWidget session={session} />
+        </ScopedCssBaseline>
+      </div>
+    </ThemeProvider>
+  )
+})
 
 export default JBrowseLinearGenomeView
