+++ products/jbrowse-react-linear-genome-view/src/createModel/AboutDialog.tsx
@@ -1,0 +1,2 @@
+// default re-export for react lazy
+export { AboutDialog as default } from '@jbrowse/product-core'
